Add response helper

+++ controllers/deleteRecipeController.js
@@ -1,11 +1,21 @@
+import { errorResponse, noContentResponse } from "../helpers/response.js";
 import Recipe from "../models/recipe.js";
 
 export default function (req, res) {
-  Recipe.findByIdAndDelete(req.params.id)
-    .then((deletedRecipe) => {
-      res.send(deletedRecipe);
+  Recipe.findById(req.params.id)
+    .then((recipe) => {
+      if (recipe === null) {
+        return noContentResponse(res);
+      }
+      Recipe.findByIdAndDelete(req.params.id)
+        .then((deletedRecipe) => {
+          res.send(deletedRecipe);
+        })
+        .catch((err) => {
+          return errorResponse(res, err);
+        });
     })
     .catch((err) => {
-      res.error(err);
+      return errorResponse(res, err);
     });
 }
